Refactored crops to accommodate fruit trees and berry bushes
