+++ src/main/java/net/dries007/tfc/objects/blocks/BlocksTFC.java
@@ -24,12 +24,8 @@ import net.minecraftforge.registries.IForgeRegistry;
 
 import net.dries007.tfc.api.registries.TFCRegistries;
 import net.dries007.tfc.api.types.*;
-import net.dries007.tfc.objects.blocks.crops.BlockCropTFC;
+import net.dries007.tfc.objects.blocks.agriculture.*;
 import net.dries007.tfc.objects.blocks.devices.*;
-import net.dries007.tfc.objects.blocks.fruittrees.BlockFruitTreeBranch;
-import net.dries007.tfc.objects.blocks.fruittrees.BlockFruitTreeLeaves;
-import net.dries007.tfc.objects.blocks.fruittrees.BlockFruitTreeSapling;
-import net.dries007.tfc.objects.blocks.fruittrees.BlockFruitTreeTrunk;
 import net.dries007.tfc.objects.blocks.metal.BlockAnvilTFC;
 import net.dries007.tfc.objects.blocks.metal.BlockIngotPile;
 import net.dries007.tfc.objects.blocks.metal.BlockMetalSheet;
